+++ src/platform_impl/linux/wayland/pointer.rs
@@ -1,3 +1,5 @@
+use std::cell::RefCell;
+use std::rc::Rc;
 use std::sync::{Arc, Mutex};
 
 use crate::event::{
@@ -17,11 +19,19 @@ use smithay_client_toolkit::reexports::protocols::unstable::relative_pointer::v1
     zwp_relative_pointer_v1::ZwpRelativePointerV1,
 };
 
+use smithay_client_toolkit::reexports::protocols::unstable::pointer_constraints::v1::client::{
+    zwp_locked_pointer_v1::ZwpLockedPointerV1, zwp_pointer_constraints_v1::Lifetime,
+    zwp_pointer_constraints_v1::ZwpPointerConstraintsV1,
+};
+
+use smithay_client_toolkit::reexports::client::protocol::wl_surface::WlSurface;
+
 pub fn implement_pointer<T: 'static>(
     seat: &wl_seat::WlSeat,
     sink: Arc<Mutex<WindowEventsSink<T>>>,
     store: Arc<Mutex<WindowStore>>,
     modifiers_tracker: Arc<Mutex<ModifiersState>>,
+    cursor_visible: Rc<RefCell<bool>>,
 ) -> WlPointer {
     seat.get_pointer(|pointer| {
         let mut mouse_focus = None;
@@ -61,6 +71,10 @@ pub fn implement_pointer<T: 'static>(
                                 },
                                 wid,
                             );
+                        }
+
+                        if *cursor_visible.borrow() == false {
+                            pointer.set_cursor(0, None, 0, 0);
                         }
                     }
                     PtrEvent::Leave { surface, .. } => {
@@ -239,5 +253,15 @@ pub fn implement_relative_pointer<T: 'static>(
             },
             (),
         )
+    })
+}
+
+pub fn implement_locked_pointer(
+    surface: &WlSurface,
+    pointer: &WlPointer,
+    constraints: &ZwpPointerConstraintsV1,
+) -> Result<ZwpLockedPointerV1, ()> {
+    constraints.lock_pointer(surface, pointer, None, Lifetime::Persistent.to_raw(), |c| {
+        c.implement_closure(|_, _| (), ())
     })
 }